Refactored MapContextSearchDrawer into TabsDrawer hierarchy.
